requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.0.3
django_coverage_plugin==1.3
Sphinx
django-extensions==1.7.4
Werkzeug==0.11.9
django-test-plus==1.0.12
factory_boy==2.7.0
django-debug-toolbar==1.4

# improved REPL
ipdb==0.10.0

pytest-django==3.1.2
pytest-sugar==0.9.1
